backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML
google-generativeai==0.3.1

# Environment and Configuration
python-dotenv==1.0.0

# CORS Support
fastapi-cors==0.0.6

# Security and Cryptography
cryptography==41.0.7

# Additional Dependencies
pydantic>=2.10.0

requests==2.32.3